back-end/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.2.20
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
lxml==5.4.0
mysqlclient==2.2.7
OWSLib==0.31.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.1
urllib3==2.2.3
